Quantum ESPRESSO files

WannierIO.guess_kpath — Method

guess_kpath(kpoints; atol)

Guess high symmetry points from kpoint coordinates.

If there is angle between two consecutive kpoints, then it is labeled as a high-symmetry point.

Arguments

kpoints: Vector of Vector or Vec3, in Cartesian coordinates

Keyword Arguments

atol: Absolute tolerance for checking cross product of two vectors

Returns

symm_point_indices: Vector of indices of high-symmetry points

symm_point_labels: Vector of labels of high-symmetry points, for the moment it is empty

WannierIO.read_qe_band — Method

read_qe_band(filename)

Read Quantum ESPRESSO bands.x output data file.

The data file has format

 &plot nbnd=  20, nks=   380 /
           -0.500000  0.500000  0.500000
   -3.320   -0.666    5.173    5.173    7.994    9.725    9.725   14.147   16.993   16.993
   17.841   17.841   17.902   19.666   25.961   26.563   28.186   28.186   28.368   28.368
           -0.495000  0.495000  0.495000
   -3.322   -0.664    5.173    5.173    7.994    9.725    9.725   14.148   16.980   16.980
...

WannierIO.read_qe_xml — Method

read_qe_xml(filename)

Read atomic structure and band structure from QE's XML output.

Return

lattice: 3 * 3, Å, each column is a lattice vector

atom_positions: length-n_atoms vector, each element is a fractional position

atom_labels: length-n_atoms vector, each element is the label of the corresponding atom

recip_lattice: 3 * 3, Å⁻¹, each column is a reciprocal lattice vector

kpoints: length-n_kpts vector, each element is a fractional kpoint

fermi_energy: eV

alat: the alat of QE in Å

eigenvalues: length-n_kpts vector, each element is a length-n_bands vector of eigenvalue in eV. For spin-polarized but without SOC calculations, return two arries of eigenvalues_up and eigenvalues_dn for the two spin channels.