Utility

Some simple convenience functions for development.

Fortran related

WannierIO.isbinary — Method

isbinary(filename::AbstractString)

Check if the file is in binary format.

WannierIO.isbinary — Method

isbinary(chars::Vector{UInt8})

Check if a sequence of chars is binary.

WannierIO.parse_bool — Method

parse_bool(s::AbstractString)

Parse a string as bool.

This is capable of parsing Fortran outputs, e.g., .true., .false., true, T.

WannierIO.parse_bool — Method

parse_bool(i::Integer)

Parse an integer as bool.

0: false

1 or -1: true

WannierIO.parse_float — Method

parse_float(s::AbstractString)

Parse a string as Float64.

The is capable of parsing Fortran outputs, e.g. 1.0D-10, to the ordinary 1e-10.

Lattice

WannierIO.get_lattice — Method

get_lattice(recip_lattice)

Compute lattice from reciprocal lattice.

WannierIO.get_recip_lattice — Method

get_recip_lattice(lattice)

Compute reciprocal lattice from lattice.

Types

WannierIO.FortranBinary — Type

Fortran unformatted IO.

WannierIO.FortranBinaryStream — Type

Fortran unformatted IO with stream access.

For example, file written using these Fortran code:

OPEN(UNIT=11, FILE="ustream.demo", STATUS="NEW", ACCESS="STREAM", FORM="UNFORMATTED")

WannierIO.FortranText — Type

Fortran formatted IO.

WannierIO.Mat3 — Type

3 x 3 matrix type.

For lattice and recip_lattice.

WannierIO.SymbolVec3 — Type

Pair type associating a Symbol with a Vec3.

For win file atoms_frac and kpoint_path.

WannierIO.Vec3 — Type

Length-3 vector type.

For atom posistions, kpoints, etc.

WannierIO.Wannier90Text — Type

Plain text format for Wannier90 win and nnkp files.

The W90 default win or nnkp are plain text files but are not simple arrays of numbers that can be read by readdlm, therefore this struct is used to indicate that the file is plain text but need to be handled by corresponding functions, e.g., read_win, read_nnkp, etc.

This somewhat overlaps with FortranText, but this one is only used for small input parameter files e.g. win and nnkp (in comparison with the Wannier90Toml format), while the FortranText (in comparison with the FortranBinary format) is used for large matrices e.g. amn, mmn, eig, etc.

WannierIO.Wannier90Toml — Type

TOML file format for Wannier90 win and nnkp files.

Here we introduce a TOML format for win and nnkp, so that once the win or nnkp files are converted into TOML, the TOML files can be loaded by standard TOML parsers without the headache of writing custom parsers in other Julia packages.

See also write_win, write_nnkp, etc.

Misc

WannierIO.default_header — Method

Default header for writing wannier90 files.

WannierIO.write_toml — Method

Write kwargs into io as a TOML file.

Do some type conversion before writing.