Wannier90 files

Warning

Some of the functions, e.g. read_amn, write_amn, read_chk, write_chk, etc., support reading/writing Fortran unformatted files. However, the Fortran unformatted files are machine and compiler dependent. Therefore, it is not guaranteed that these functions work for all the cases. Currently, the functions are tested on the following platforms:

Linux + gfortran 11.2

Contents

Wannier90 files

Contents

Index

Read/write

Index

WannierIO.Chk

Base.isapprox

WannierIO._fix_win!

WannierIO._read_amn_bin

WannierIO._read_amn_fmt

WannierIO._read_chk_bin

WannierIO._read_chk_fmt

WannierIO._read_eig_bin

WannierIO._read_eig_fmt

WannierIO._read_mmn_bin

WannierIO._read_mmn_fmt

WannierIO._read_nnkp_toml

WannierIO._read_unk_bin

WannierIO._read_unk_fmt

WannierIO._read_win_toml

WannierIO._reshape_eig

WannierIO._write_amn_bin

WannierIO._write_amn_fmt

WannierIO._write_chk_bin

WannierIO._write_chk_fmt

WannierIO._write_eig_bin

WannierIO._write_eig_fmt

WannierIO._write_mmn_bin

WannierIO._write_mmn_fmt

WannierIO._write_nnkp_toml

WannierIO._write_unk_bin

WannierIO._write_unk_fmt

WannierIO._write_win_toml

WannierIO.get_U

WannierIO.get_Udis

WannierIO.read_amn

WannierIO.read_chk

WannierIO.read_eig

WannierIO.read_mmn

WannierIO.read_nnkp

WannierIO.read_spn

WannierIO.read_unk

WannierIO.read_w90_band

WannierIO.read_w90_band_dat

WannierIO.read_w90_band_kpt

WannierIO.read_w90_band_labelinfo

WannierIO.read_w90_tbdat

WannierIO.read_w90_wsvec

WannierIO.read_win

WannierIO.read_wout

WannierIO.write_HH_R

WannierIO.write_amn

WannierIO.write_amn

WannierIO.write_chk

WannierIO.write_eig

WannierIO.write_mmn

WannierIO.write_nnkp

WannierIO.write_spn

WannierIO.write_spn

WannierIO.write_unk

WannierIO.write_w90_band

WannierIO.write_w90_band_dat

WannierIO.write_w90_band_kpt

WannierIO.write_w90_band_labelinfo

WannierIO.write_win

Read/write

WannierIO._fix_win! — Method

_fix_win!(params)

Sanity check and add missing input parameters from a win file.

See also read_win.

WannierIO._read_win_toml — Method

Read a win file in TOML format.

Used only in tests.

WannierIO._write_win_toml — Method

Write a win file in TOML format.

Used only in tests. Currently the output file is not formatted nicely.

WannierIO.read_win — Method

read_win(filename; fix_inputs=true)

Read the input file of wannier90.

Arguments

filename: The name of the input file.

Keyword Arguments

fix_inputs: sanity check and fix the input parameters, e.g., set num_bands = num_wann if num_bands is not specified, convert atoms_cart always to atoms_frac, etc. See also _fix_win!.

WannierIO.write_win — Method

write_win(filename; kwargs...)

Write input parameters into a wannier90 win file.

The input parameters are keyword arguments, with key names same as that of wannier90.

Examples

using WannierIO

write_win(
    "silicon.win";
    num_wann=4,
    num_bands=4,
    # unit_cell_cart is a matrix, its columns are the lattice vectors in angstrom
    unit_cell_cart=[
        0.0      2.71527  2.71527
        2.71527  0.0      2.71527
        2.71527  2.71527  0.0
    ],
    # atoms_frac is a vector of pairs of atom_label and fractional coordinates
    atoms_frac=[
        :Si => [0.0, 0.0, 0.0],
        :Si => [0.25, 0.25, 0.25],
        # both `:Si` and `"Si"` are allowed
        # "Si" => [0.25, 0.25, 0.25],
    ],
    # each element in projections will be written as a line in the win file
    projections=[
        "random",
    ]
    kpoint_path=[
        [:G => [0.0, 0.0, 0.0], :X => [0.5, 0.0, 0.5]],
        [:X => [0.5, 0.0, 0.5], :U => [0.625, 0.25, 0.625]],
    ],
    mp_grid=[2, 2, 2],
    # kpoints is a matrix, its columns are the fractional coordinates
    kpoints=[
        [0.0, 0.0, 0.0],
        [0.0, 0.0, 0.5],
        [0.0, 0.5, 0.0],
        [0.0, 0.5, 0.5],
        [0.5, 0.0, 0.0],
        [0.5, 0.0, 0.5],
        [0.5, 0.5, 0.0],
        [0.5, 0.5, 0.5],
    ],
    # additional parameters can be passed as keyword arguments, e.g.,
    num_iter=500,
)

WannierIO.read_wout — Method

read_wout(filename::AbstractString)

Parse wout file.

Return

lattice: in Å, each column is a lattice vector

atom_labels: atomic symbols

atom_positions: in fractional coordinates

centers: WF centers in Å

spreads: WF spreads in Å^2

WannierIO._read_amn_bin — Method

_read_amn_bin(filename::AbstractString)

Read Fortran binary (unformatted) amn file.

Using Fortran stream IO.

WannierIO._read_amn_fmt — Method

_read_amn_fmt(filename::AbstractString)

Read Fortran formatted amn file.

WannierIO._write_amn_bin — Method

Write A as Fortran unformatted file.

WannierIO._write_amn_fmt — Method

Write A in plain text format.

WannierIO.read_amn — Method

read_amn(filename::AbstractString)

Read the amn file.

Return

A: length-n_kpts vector, each element is a n_bands * n_wann matrix.

WannierIO.write_amn — Method

write_amn(filename, A, header; binary=false)

Write amn file.

Arguments

filename: output filename

A: a length-n_kpts vector, each element is a n_bands * n_wann matrix

header: 1st line of the file

Keyword arguments

binary: write as Fortran unformatted file

WannierIO.write_amn — Method

write_amn(filename, A)

Write amn file.

Default header is Created by WannierIO.jl CURRENT_DATE.

Arguments

filename: output filename

A: a length-n_kpts vector, each element is a n_bands * n_wann matrix

Keyword arguments

binary: write as Fortran unformatted file

WannierIO._read_mmn_bin — Method

Read Fortran binary mmn file.

WannierIO._read_mmn_fmt — Method

Read plain text mmn file.

WannierIO._write_mmn_bin — Method

Write binary mmn file.

WannierIO._write_mmn_fmt — Method

Write plain text mmn file.

WannierIO.read_mmn — Method

read_mmn(filename::AbstractString)

Read mmn file.

Return

M: length-n_kpts vector of n_bands * n_bands * n_bvecs arrays

kpb_k: length-n_kpts vector of length-n_bvecs vector of integers

kpb_b: length-n_kpts vector of length-n_bvecs vector of Vec3{Int} for bvectors

WannierIO.write_mmn — Function

write_mmn(filename, M, kpb_k, kpb_b, header; binary=false)

Write mmn file.

Arguments

filename: output file name

M: length-n_kpts vector of n_bands * n_bands * n_bvecs arrays

kpb_k: length-n_kpts vector of length-n_bvecs vector of integers

kpb_b: length-n_kpts vector of length-n_bvecs vector of Vec3{Int} for bvectors

header: header string

Keyword arguments

binary: if true write in Fortran binary format

WannierIO._read_eig_bin — Method

Read binary eig file.

WannierIO._read_eig_fmt — Method

Read plain text eig file.

WannierIO._reshape_eig — Method

Reshape a vector of eigenvalues into a matrix of eigenvalues.

WannierIO._write_eig_bin — Method

Write binary eig file.

WannierIO._write_eig_fmt — Method

Write plain text eig file.

WannierIO.read_eig — Method

read_eig(filename::AbstractString)

Read the eig file.

Return

E: a lenth-n_kpts vector, each element is a length-n_bands vector

WannierIO.write_eig — Method

write_eig(filename::AbstractString, E::AbstractArray; binary=false)

Write eig file.

Arguments

E: a lenth-n_kpts vector, each element is a length-n_bands vector

Keyword arguments

binary: if true write in Fortran binary format

WannierIO.read_spn — Method

read_spn(filename::AbstractString)

Read the spn file.

Return a n_bands * n_bands * n_kpts * 3 array.

WannierIO.write_spn — Method

write_spn(filename::String, S::Array{Complex,4}, header::String)

Write the spn file.

WannierIO.write_spn — Method

write_spn(filename::String, S::Array{Complex,4})

Write the spn file.

The header is "Created by WannierIO.jl CURRENT_DATE".

WannierIO._read_unk_bin — Method

Read binary unk file.

WannierIO._read_unk_fmt — Method

Read plain text unk file.

WannierIO._write_unk_bin — Method

Write binary unk file.

WannierIO._write_unk_fmt — Method

Write plain text unk file.

WannierIO.read_unk — Method

read_unk(filename::AbstractString)

Read UNK file for the periodic part of Bloch wavefunctions.

Return

ik: k-point index

Ψ: periodic part of Bloch wavefunctions in real space, size = (n_gx, n_gy, n_gz, n_bands, n_spin)

WannierIO.write_unk — Method

write_unk(filename::AbstractString, ik::Integer, Ψ::Array{Complex,4}; binary=false)

Write UNK file for the periodic part of Bloch wavefunctions.

Arguments

ik: at which kpoint? start from 1

Ψ: Bloch wavefunctions, size(Ψ) = (n_gx, n_gy, n_gz, n_bands, n_spin)

Keyword arguments

binary: write as Fortran unformatted file

WannierIO._read_nnkp_toml — Method

Read a nnkp file in TOML format.

Used only in tests.

WannierIO._write_nnkp_toml — Method

Write a nnkp file in TOML format.

Used only in tests. Internally calls _write_win_toml.

WannierIO.read_nnkp — Method

read_nnkp(filename::AbstractString)

Read the nnkp file.

Return

recip_lattice: each column is a reciprocal lattice vector

kpoints: length-n_kpts vector of Vec3, in fractional coordinates

kpb_k: length-n_kpts vector of length-n_bvecs vector of integers, index of kpoints

kpb_b: length-n_kpts vector of length-n_bvecs vector of Vec3 for bvectors, fractional w.r.t recip_lattice

WannierIO.write_nnkp — Function

write_nnkp(filename, bvectors::BVectors, n_wann)

Write a nnkp file that can be used by DFT codes, e.g., QE pw2wannier90.

Arguments

recip_lattice: each column is a reciprocal lattice vector

kpoints: length-n_kpts vector of Vec3, in fractional coordinates

kpb_k: length-n_kpts vector of length-n_bvecs vector of integers, index of kpoints

kpb_b: length-n_kpts vector of length-n_bvecs vector of Vec3 for bvectors, fractional w.r.t recip_lattice

n_wann: if given, write a auto_projections block

exclude_bands: if given, write the specified band indexes in the exclude_bands block

Note

Only a preliminary version, use auto_projections.

WannierIO.Chk — Type

Struct for storing matrices in seedname.chk file.

Base.isapprox — Method

Compare two Chk objects.

Used in tests.

WannierIO._read_chk_bin — Method

Read unformatted seedname.chk file.

WannierIO._read_chk_fmt — Method

Read formatted seedname.chk file.

WannierIO._write_chk_bin — Method

Write unformatted chk file.

WannierIO._write_chk_fmt — Method

Write formatted chk file.

WannierIO.get_U — Method

get_U(chk)

Extract U matrices from Chk.

WannierIO.get_Udis — Method

get_Udis(chk)

Extract U matrices for disentanglement from Chk.

WannierIO.read_chk — Method

read_chk(filename::AbstractString)

Read formatted or binary chk file.

WannierIO.write_chk — Method

write_chk(filename, chk::Chk)

Write formatted or binary chk file.

Keyword arguments

binary: write as Fortran unformatted file

WannierIO.read_w90_tbdat — Method

read_w90_tbdat(filename)

Read seedname_tb.dat.

Return

lattice: each column is a lattice vector

R: $\bm{R}$ vectors on which operators are defined

N: degeneracies of $\bm{R}$ vectors

H: Hamiltonian $\bm{H}(\bm{R})$

r: position operator

WannierIO.read_w90_wsvec — Method

read_w90_wsvec(filename::AbstractString)

Read seedname_wsvec.dat.

WannierIO.write_HH_R — Method

write_HH_R(filename, H, R; N=nothing, header=nothing)

Write the real space Hamiltonian to a seedname_HH_R.dat file.

Arguments

filename: usually seedname_HH_R.dat

H: a n_wann * n_wann * n_rvecs array of Hamiltonian

R: a n_rvecs * 3 array of integers

Keyword arguments

N: a n_rvecs vector of integers, the degeneracy of each R vector

header: a string, the header of the file

Note

Wanier90 postw90.x has a hidden input parameter effective_model, setting it to true and postw90.x will read this HH_R.dat to fill the real space Hamiltonian, and do subsequent Wannier interpolation, e.g., in BoltzWann. However, the vanila postw90.x code does not take into account the degeneracy of R vectors, and also does not use MDRS interpolation. I have modified the postw90.x code to use MDRS, and also changed a bit the number of digits for the Hamiltonian in HH_R.dat, so that it is the same as the seedname_tb.dat file, i.e., from Fortran F12.6 to E15.8.

WannierIO.read_w90_band — Method

read_w90_band(seedname::AbstractString)

Read SEEDNAME_band.dat, SEEDNAME_band.kpt, SEEDNAME_band.labelinfo.dat.

Return

kpoints: 3 * n_kpts, fractional coordinates

E: length-n_kpts vector, each element is a length-n_bands vector of band energies

x: n_kpts, x axis value, in cartesian length

symm_idx: index of high-symmetry points in seedname_band.dat

symm_label: name of high-symmetry points

WannierIO.read_w90_band_dat — Method

read_w90_band_dat(filename)

Read seedname_band.dat file.

Return

x: n_kpts, x axis value, in cartesian length

E: length-n_kpts vector, each elemnt is a length-n_bands vector of band energies

WannierIO.read_w90_band_kpt — Method

read_w90_band_kpt(filename)

Read seedname_band.kpt file.

Return

kpoints: a vector of length n_kpts, fractional coordinates

weights: a vector of length n_kpts, weights of kpoints

WannierIO.read_w90_band_labelinfo — Method

read_w90_band_labelinfo(filename::AbstractString)

Read seedname_band.labelinfo file.

Return

symm_idx: index of high-symmetry points in seedname_band.dat

symm_label: name of high-symmetry points

WannierIO.write_w90_band — Method

write_w90_band(seedname, kpoints, E, x, symm_idx, symm_label)

Write SEEDNAME_band.dat, SEEDNAME_band.kpt, SEEDNAME_band.labelinfo.dat.

Arguments

seedname: seedname of SEEDNAME_band.dat, SEEDNAME_band.kpt, SEEDNAME_band.labelinfo.dat

kpoints: length-n_kpts vector, fractional coordinates

E: length-n_kpts vector, each element is a length-n_bands vector of band energies

x: n_kpts, x axis value, in cartesian length

symm_idx: index of high-symmetry points in seedname_band.dat

symm_label: name of high-symmetry points

WannierIO.write_w90_band_dat — Method

write_w90_band_dat(filename, x, E)

Write seedname_band.dat file.

Arguments

filename: filename of seedname_band.dat

x: n_kpts, x axis value, in cartesian length

E: length-n_kpts vector, each element is a length-n_bands vector of band energies

WannierIO.write_w90_band_kpt — Function

write_w90_band_kpt(filename, kpoints, weights=nothing)

Write seedname_band.kpt file.

Arguments

filename: filename of seedname_band.kpt

kpoints: length-n_kpts vector, fractional coordinates

weights: n_kpts, optional, weights of kpoints

WannierIO.write_w90_band_labelinfo — Method

write_w90_band_labelinfo(filename, symm_idx, symm_label, x, kpoints)

Write seedname_band.labelinfo file.

Arguments

filename: filename of seedname_band.labelinfo

symm_idx: index of high-symmetry points in seedname_band.dat

symm_label: name of high-symmetry points

x: n_kpts, x axis value, in cartesian length

kpoints: length-n_kpts vector, fractional coordinates